Remove external dependencies for command line client.
