+++ cli/mobilecoin/__init__.py
@@ -2,8 +2,6 @@ from mobilecoin.cli import CommandLineInterface
 from mobilecoin.client import (
     Client,
     WalletAPIError,
-)
-from mobilecoin.utility import (
     mob2pmob,
     pmob2mob,
 )
