+++ cli/setup.py
@@ -11,9 +11,5 @@ setup(
     url='https://github.com/christian-oudard/mobilecoin-python-cli',
     packages=['mobilecoin'],
     scripts=['bin/mobcli'],
-    install_requires=[
-        'requests',
-        'base58',
-        'protobuf',
-    ],
+    install_requires=[],
 )
